requirements.txt
altair==5.4.1
annotated-types==0.7.0
attrs==24.2.0
blinker==1.8.2
blis==1.0.1
cachetools==5.5.0
catalogue==2.0.10
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
cloudpathlib==0.20.0
colorama==0.4.6
confection==0.1.5
contourpy==1.1.1
cycler==0.12.1
cymem==2.0.8
filelock==3.16.1
fonttools==4.55.0
fsspec==2024.10.0
gitdb==4.0.11
GitPython==3.1.43
huggingface-hub==0.26.2
idna==3.10
importlib_resources==6.4.5
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.7
langcodes==3.5.0
language_data==1.3.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.7.5
mdurl==0.1.2
murmurhash==1.0.10
narwhals==1.14.1
nltk==3.9.1
numpy==1.26.4
packaging==24.2
pandas==2.2.3
pdf2image==1.17.0
pillow==10.4.0
pkgutil_resolve_name==1.3.10
preshed==3.0.9
protobuf==5.28.3
pyarrow==17.0.0
pydantic==2.10.0
pydantic_core==2.27.0
pydeck==0.9.1
Pygments==2.18.0
pyparsing==3.1.4
pytesseract==0.3.13
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rpds-py==0.20.1
safetensors==0.4.5
seaborn==0.13.2
setuptools==75.6.0
shellingham==1.5.4
six==1.16.0
smart-open==7.0.5
smmap==5.0.1
spacy==3.8.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
streamlit==1.40.1
tenacity==9.0.0
thinc==8.3.2
tokenizers==0.20.3
toml==0.10.2
tornado==6.4.1
tqdm==4.67.0
transformers==4.46.3
typer==0.13.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
wasabi==1.1.3
watchdog==4.0.2
weasel==0.4.1
wrapt==1.16.0
zipp==3.20.2
